change servos

+++ src/pbot_servos/src/servos_wrapper.py
@@ -2,7 +2,7 @@
 
 import servo
 import rospy
-from std_msgs.msg import Int8
+from std_msgs.msg import Int16
 
 
 class ServosWrapper:
@@ -11,14 +11,14 @@ class ServosWrapper:
         self._current_cam_h_pos = 90
 
         rospy.init_node('pbot_servos')
-        rospy.Subscriber("/pbot/cam_h_servo/target_pos", Int8, self.callback_cam_h)
+        rospy.Subscriber("/pbot/cam_h_servo/target_pos", Int16 , self.callback_cam_h)
 
-        self.pub_h_servo = rospy.Publisher("/pbot/cam_h_servo/current_pos", Int8, queue_size=10)
+        self.pub_h_servo = rospy.Publisher("/pbot/cam_h_servo/current_pos", Int16, queue_size=10)
 
     def cleanup(self):
         del self.cam_h_servo
 
-    def callback_cam_h(self, msg: Int8):
+    def callback_cam_h(self, msg: Int16):
         rospy.loginfo("cam_h. Target position %s", msg.data)
 
         self._current_cam_h_pos = self.cam_h_servo.servo_appointed_detection(msg.data)
